requirements.txt
streamlit
google-generativeai
fpdf2
python-dotenv